requirements.txt
aiofiles==23.2.1
aiogram==3.6.0
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.7.0
asgiref==3.8.1
async-timeout==4.0.3
attrs==23.2.0
black==24.3.0
certifi==2024.2.2
click==8.1.7
Django==5.0.3
django-jazzmin==2.6.1
django-modeltranslation==0.18.13
djangorestframework==3.15.1
drf-spectacular==0.27.1
drf-spectacular-sidecar==2024.3.4
environs==11.0.0
frozenlist==1.4.1
gunicorn==21.2.0
idna==3.7
inflection==0.5.1
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
magic-filter==1.0.12
marshmallow==3.21.2
multidict==6.0.5
mypy-extensions==1.0.0
packaging==24.0
pathspec==0.12.1
pillow==10.2.0
platformdirs==4.2.2
psycopg2-binary==2.9.9
pydantic==2.7.1
pydantic_core==2.18.2
python-dotenv==1.0.1
PyYAML==6.0.1
referencing==0.35.1
rpds-py==0.18.1
sqlparse==0.5.0
tomli==2.0.1
typing_extensions==4.11.0
uritemplate==4.1.1
whitenoise==6.6.0
yarl==1.9.4
